refactor: restructure useCardTags hook
