+++ frontend/src/pages/settings/card-display.tsx
@@ -2,12 +2,12 @@ import type { Settings } from "@arkham-build/shared";
 import { useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { CardTagList } from "@/components/card-tags/card-tag-list";
-import type { TagItem } from "@/components/card-tags/use-card-tags";
 import { ListCardInner } from "@/components/list-card/list-card-inner";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Field, FieldLabel } from "@/components/ui/field";
 import { Select } from "@/components/ui/select";
 import { useStore } from "@/store";
+import type { TagItem } from "@/store/selectors/card-tags";
 import { selectMetadata } from "@/store/selectors/shared";
 import css from "./card-display.module.css";
 import type { SettingProps } from "./types";
